// gme-web/gme-web/src/app/@data/model/SolicitudExterna.ts
export class SolicitudExterna {
  abreviatura: any;
  anexo: any;
  apellidoMaterno: any;
  apellidoPaterno: any;
  celular: any;
  correoElectronico: any;
  estado: any;
  estadoId: number;
  fechaNacimiento: any;
  fechaSolicitud: any;
  nivelGobierno: any;
  nivelGobiernoId: number;
  nombreCompleto: any;
  nombreEntidad: any;
  nombres: any;
  numeroDocumento: any;
  razonSocial: any;
  rucEntidad: any;
  sector: any;
  sectorId: number;
  solicitudEntidadExtId: number;
  telefonoFijo: any;
  tipoDocumento: number;
  tipoEntidad: any;
  tipoEntidadId: number;
  urlLogoEntidad: any;
  uuidDocumento: any;
  colorEstado: any;
  correoGestorGdr: any;
  sindicato: any;

}